refactor(character): change import path
